+++ connector_search_engine/tests/test_all.py
@@ -176,7 +176,7 @@ class TestBindingIndex(TestBindingIndexBaseFake):
     def test_recompute_all_indexes(self):
         # on creation indexes are computed and external data stored
         expected = {
-            "id": self.partner_binding.id,
+            "id": self.partner_binding.record_id.id,
             "active": True,
             "lang": "en_US",
             "name": "Marty McFly",
@@ -360,3 +360,17 @@ class TestBindingIndex(TestBindingIndexBaseFake):
         result = self.partner_binding.recompute_json()
         self.assertEqual(self.partner_binding.sync_state, "to_update")
         self.assertEqual(result, "")
+
+    def test_customize_id_key_without_target(self):
+        self.env["ir.exports.line"].create(
+            {"export_id": self.exporter.id, "name": "id"}
+        )
+        self.partner_binding.recompute_json()
+        self.assertEqual(self.partner_binding.data["id"], self.partner_binding.id)
+
+    def test_customize_id_key_with_target(self):
+        self.env["ir.exports.line"].create(
+            {"export_id": self.exporter.id, "name": "name:id"}
+        )
+        self.partner_binding.recompute_json()
+        self.assertEqual(self.partner_binding.data["id"], "Marty McFly")